add check for id

+++ src/store/modules/user/index.js
@@ -65,7 +65,7 @@ const actions = {
   },
   createUser (payload) {
     return new Promise((resolve, reject) => {
-      axios.post(server + '/api/user/' + payload)
+      axios.post(server + '/api/user/', payload)
         .then(function (responce) {
         })
         .catch(function (err) {
@@ -75,26 +75,34 @@ const actions = {
     })
   },
   updateUser (payload) {
-    return new Promise((resolve, reject) => {
-      axios.post(server + '/api/user/:' + payload.username)
-        .then(function (responce) {
-        })
-        .catch(function (err) {
-          reject(err)
-          console.log(err)
-        })
-    })
+    const id = state.user.id
+
+    if (id) {
+      return new Promise((resolve, reject) => {
+        axios.post(server + '/api/user/:' + payload.username)
+          .then(function (responce) {
+          })
+          .catch(function (err) {
+            reject(err)
+            console.log(err)
+          })
+      })
+    }
   },
   deleteUser (payload) {
-    return new Promise((resolve, reject) => {
-      axios.delete(server + '/api/user/:' + payload.username)
-        .then(function (responce) {
-        })
-        .catch(function (err) {
-          reject(err)
-          console.log(err)
-        })
-    })
+    const id = state.user.id
+
+    if (id) {
+      return new Promise((resolve, reject) => {
+        axios.delete(server + '/api/user/:' + payload.username)
+          .then(function (responce) {
+          })
+          .catch(function (err) {
+            reject(err)
+            console.log(err)
+          })
+      })
+    }
   }
 }
